requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
yt-dlp
torch
torchaudio
python-dotenv
requests